+++ rouver/handler.py
@@ -1,6 +1,6 @@
 import collections
 from http import HTTPStatus
-from typing import cast, Any, Union, Iterator, Sequence, Dict, Iterable
+from typing import cast, Any, Union, Iterator, Sequence, Dict, Iterable, List
 
 from werkzeug.wrappers import Request
 
@@ -33,15 +33,21 @@ class RouteHandlerBase(collections.Iterable):
     ...         ])
     """
 
-    def __init__(self, request: Request, path_args: Sequence[Any],
+    def __init__(self, request: Request, _: Sequence[Any],
                  start_response: StartResponse) -> None:
         self.request = request
-        self.path_args = path_args
         self.start_response = start_response
         self._response = self.prepare_response()
 
     def __iter__(self) -> Iterator[bytes]:
         return iter(self._response)
+
+    @property
+    def path_args(self) -> List[Any]:
+        path_args = self.request.environ.get("rouver.path_args")
+        if not isinstance(path_args, list):
+            return []
+        return path_args
 
     def prepare_response(self) -> Iterable[bytes]:
         raise NotImplementedError()